// src/api/generated/models/issuePriority.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export type IssuePriority = typeof IssuePriority[keyof typeof IssuePriority];


export const IssuePriority = {
  no_priority: 'no_priority',
  urgent: 'urgent',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;